// src/generated/prisma/enums.ts
export const MODULES = {
  ALL: 'ALL',
  HOME: 'HOME',
  REPORTABILITY: 'REPORTABILITY',
  REPORTABILITY_OTC: 'REPORTABILITY_OTC',
  NOTIFICATIONS: 'NOTIFICATIONS',
  TOOLS: 'TOOLS',
  TUTORIALS: 'TUTORIALS',
  EQUIPMENT: 'EQUIPMENT',
  EQUIPMENT_HISTORY: 'EQUIPMENT_HISTORY',
  SAFETY_TALK: 'SAFETY_TALK',
  WORK_ORDERS: 'WORK_ORDERS',
  INDICATORS: 'INDICATORS',
  WORK_PERMITS: 'WORK_PERMITS',
  LOCKOUT_PERMITS: 'LOCKOUT_PERMITS',
  DOCUMENTATION: 'DOCUMENTATION',
  WORK_REQUESTS: 'WORK_REQUESTS',
  COMPANY: 'COMPANY',
  USERS: 'USERS',
  MAINTENANCE_PLANS: 'MAINTENANCE_PLANS',
  PROGRAMATION: 'PROGRAMATION',
  STARTUP_FOLDERS: 'STARTUP_FOLDERS',
  LABOR_CONTROL_FOLDERS: 'LABOR_CONTROL_FOLDERS',
  VEHICLES: 'VEHICLES',
  CONTACT: 'CONTACT',
  NONE: 'NONE'
} as const

export type MODULES = (typeof MODULES)[keyof typeof MODULES]
// ...
export const ReviewStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  EXPIRED: 'EXPIRED',
  TO_UPDATE: 'TO_UPDATE',
  NOT_APPLIED: 'NOT_APPLIED'
} as const

export type ReviewStatus = (typeof ReviewStatus)[keyof typeof ReviewStatus]


export const SafetyAndHealthDocumentType = {
  COMPANY_INFO: 'COMPANY_INFO',
  STAFF_LIST: 'STAFF_LIST',
  GANTT_CHART: 'GANTT_CHART',
  MUTUAL: 'MUTUAL',
  INTERNAL_REGULATION: 'INTERNAL_REGULATION',
  ACCIDENT_RATE: 'ACCIDENT_RATE',
  RISK_MATRIX: 'RISK_MATRIX',
  PREVENTION_PLAN: 'PREVENTION_PLAN',
  WORK_PROCEDURE: 'WORK_PROCEDURE',
  EMERGENCY_PROCEDURE: 'EMERGENCY_PROCEDURE',
  TOOLS_MAINTENANCE: 'TOOLS_MAINTENANCE',
  PPE_CERTIFICATION: 'PPE_CERTIFICATION',
  HARASSMENT_PROCEDURE: 'HARASSMENT_PROCEDURE',
  ORGANIZATION_CHART: 'ORGANIZATION_CHART',
  SAFE_WORK: 'SAFE_WORK',
  RISK_ANALYSIS: 'RISK_ANALYSIS',
  WORK_PERMIT: 'WORK_PERMIT',
  CONTRACTOR_REGULATION_RECEIPT: 'CONTRACTOR_REGULATION_RECEIPT'
} as const

export type SafetyAndHealthDocumentType = (typeof SafetyAndHealthDocumentType)[keyof typeof SafetyAndHealthDocumentType]


export const WorkerDocumentType = {
  CONTRACT: 'CONTRACT',
  INTERNAL_REGULATION_RECEIPT: 'INTERNAL_REGULATION_RECEIPT',
  RISK_INFORMATION: 'RISK_INFORMATION',
  ID_CARD: 'ID_CARD',
  DRIVING_LICENSE: 'DRIVING_LICENSE',
  HEALTH_EXAM: 'HEALTH_EXAM',
  PSYCHOTECHNICAL_EXAM: 'PSYCHOTECHNICAL_EXAM',
  RISK_MATRIX_TRAINING: 'RISK_MATRIX_TRAINING',
  WORK_PROCEDURE_TRAINING: 'WORK_PROCEDURE_TRAINING',
  EMERGENCY_PROCEDURE_TRAINING: 'EMERGENCY_PROCEDURE_TRAINING',
  DEFENSIVE_DRIVING_TRAINING: 'DEFENSIVE_DRIVING_TRAINING',
  MOUNTAIN_DEFENSIVE_DRIVING: 'MOUNTAIN_DEFENSIVE_DRIVING',
  TOOLS_MAINTENANCE_TRAINING: 'TOOLS_MAINTENANCE_TRAINING',
  HARASSMENT_TRAINING: 'HARASSMENT_TRAINING',
  PPE_RECEIPT: 'PPE_RECEIPT',
  PREVENTION_EXPERT: 'PREVENTION_EXPERT',
  HIGH_RISK_TRAINING: 'HIGH_RISK_TRAINING',
  ENVIRONMENTAL_TRAINING: 'ENVIRONMENTAL_TRAINING',
  ALCOHOL_AND_DRUGS_EXAM: 'ALCOHOL_AND_DRUGS_EXAM',
  IRL_SAFETY_TALK: 'IRL_SAFETY_TALK'
} as const

export type WorkerDocumentType = (typeof WorkerDocumentType)[keyof typeof WorkerDocumentType]


export const VehicleDocumentType = {
  EQUIPMENT_FILE: 'EQUIPMENT_FILE',
  VEHICLE_REGISTRATION: 'VEHICLE_REGISTRATION',
  CIRCULATION_PERMIT: 'CIRCULATION_PERMIT',
  TECHNICAL_REVIEW: 'TECHNICAL_REVIEW',
  INSURANCE: 'INSURANCE',
  CHECKLIST: 'CHECKLIST',
  TRANSPORTATION_TO_OTC: 'TRANSPORTATION_TO_OTC',
  HAZARDOUS_WASTE_TRANSPORT: 'HAZARDOUS_WASTE_TRANSPORT',
  NON_HAZARDOUS_WASTE_TRANSPORT: 'NON_HAZARDOUS_WASTE_TRANSPORT',
  LIQUID_WASTE_TRANSPORT: 'LIQUID_WASTE_TRANSPORT'
} as const

export type VehicleDocumentType = (typeof VehicleDocumentType)[keyof typeof VehicleDocumentType]


export const EnvironmentalDocType = {
  ENVIRONMENTAL_PLAN: 'ENVIRONMENTAL_PLAN',
  SPILL_PREVENTION: 'SPILL_PREVENTION',
  WASTE_MANAGEMENT: 'WASTE_MANAGEMENT',
  ENVIRONMENTAL_TRAINING: 'ENVIRONMENTAL_TRAINING',
  ENVIRONMENTAL_MATRIX: 'ENVIRONMENTAL_MATRIX',
  RECT_CERTIFICATE: 'RECT_CERTIFICATE',
  WATER_CERTIFICATE: 'WATER_CERTIFICATE',
  WATER_FACTORY_RESOLUTION: 'WATER_FACTORY_RESOLUTION',
  DINING_RESOLUTION: 'DINING_RESOLUTION',
  CHEMICAL_TOILET_CONTRACT: 'CHEMICAL_TOILET_CONTRACT',
  SAFETY_DATA_SHEET: 'SAFETY_DATA_SHEET',
  LAYOUT_PLAN: 'LAYOUT_PLAN',
  ELECTRICAL_DECLARATION: 'ELECTRICAL_DECLARATION',
  GAS_DECLARATION: 'GAS_DECLARATION',
  PEST_CONTROL_RESOLUTION: 'PEST_CONTROL_RESOLUTION',
  PEST_CONTROL_CERTIFICATE: 'PEST_CONTROL_CERTIFICATE',
  PEST_CONTROL_PRODUCTS_SDS: 'PEST_CONTROL_PRODUCTS_SDS',
  PEST_CONTROL_TRACKING: 'PEST_CONTROL_TRACKING',
  HAZARDOUS_STORAGE_CHECKLIST: 'HAZARDOUS_STORAGE_CHECKLIST',
  HAZARDOUS_INVENTORY: 'HAZARDOUS_INVENTORY',
  LAND_MOVEMENT_PERMIT: 'LAND_MOVEMENT_PERMIT',
  DEBRIS_ROUTE: 'DEBRIS_ROUTE',
  FUEL_CONSUMPTION: 'FUEL_CONSUMPTION',
  WATER_CONSUMPTION: 'WATER_CONSUMPTION',
  OTHER: 'OTHER'
} as const

export type EnvironmentalDocType = (typeof EnvironmentalDocType)[keyof typeof EnvironmentalDocType]


export const EnvironmentDocType = {
  WORK_PROCEDURE: 'WORK_PROCEDURE',
  ENVIRONMENTAL_ASPECTS_AND_IMPACTS_MATRIX: 'ENVIRONMENTAL_ASPECTS_AND_IMPACTS_MATRIX',
  SAFETY_DATA_SHEET_FOR_CHEMICALS: 'SAFETY_DATA_SHEET_FOR_CHEMICALS',
  WORKER_TRAINING_RECORD: 'WORKER_TRAINING_RECORD',
  HEALTH_RESOLUTION_FOR_WORKERS_DRINKING_WATER: 'HEALTH_RESOLUTION_FOR_WORKERS_DRINKING_WATER',
  HEALTH_RESOLUTION_FOR_THE_CHEMICAL_TOILET: 'HEALTH_RESOLUTION_FOR_THE_CHEMICAL_TOILET',
  RESOLUTION_FOR_THE_SITE_WHERE_DEBRIS_WILL_BE_DISPOSED: 'RESOLUTION_FOR_THE_SITE_WHERE_DEBRIS_WILL_BE_DISPOSED',
  RESOLUTION_FOR_THE_DEBRIS_TRANSPORTER: 'RESOLUTION_FOR_THE_DEBRIS_TRANSPORTER',
  DEBRIS_TRANSFER_ROUTE: 'DEBRIS_TRANSFER_ROUTE',
  HEALTH_RESOLUTION_FROM_THE_PEST_CONTROL_COMPANY: 'HEALTH_RESOLUTION_FROM_THE_PEST_CONTROL_COMPANY',
  ENVIRONMENTAL_MANAGEMENT_PLAN: 'ENVIRONMENTAL_MANAGEMENT_PLAN'
} as const

export type EnvironmentDocType = (typeof EnvironmentDocType)[keyof typeof EnvironmentDocType]


export const BasicDocumentType = {
  CONTRACT: 'CONTRACT',
  INSURANCE: 'INSURANCE',
  PPE_RECEIPT: 'PPE_RECEIPT',
  SAFETY_AND_HEALTH_INFO: 'SAFETY_AND_HEALTH_INFO',
  IRL_SAFETY_TALK: 'IRL_SAFETY_TALK'
} as const

export type BasicDocumentType = (typeof BasicDocumentType)[keyof typeof BasicDocumentType]


export const TechSpecsDocumentType = {
  GANTT_CHART: 'GANTT_CHART',
  TECHNICAL_WORK_PROCEDURE: 'TECHNICAL_WORK_PROCEDURE'
} as const

export type TechSpecsDocumentType = (typeof TechSpecsDocumentType)[keyof typeof TechSpecsDocumentType]


export const DocumentCategory = {
  VEHICLES: 'VEHICLES',
  PERSONNEL: 'PERSONNEL',
  ENVIRONMENTAL: 'ENVIRONMENTAL',
  ENVIRONMENT: 'ENVIRONMENT',
  SAFETY_AND_HEALTH: 'SAFETY_AND_HEALTH',
  BASIC: 'BASIC',
  TECHNICAL_SPECS: 'TECHNICAL_SPECS'
} as const

export type DocumentCategory = (typeof DocumentCategory)[keyof typeof DocumentCategory]
// ...
export const ACTIVITY_TYPE = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
  SUBMIT: 'SUBMIT',
  APPROVE: 'APPROVE',
  REJECT: 'REJECT',
  CANCEL: 'CANCEL',
  COMPLETE: 'COMPLETE',
  VIEW: 'VIEW',
  DOWNLOAD: 'DOWNLOAD',
  UPLOAD: 'UPLOAD',
  COMMENT: 'COMMENT',
  ASSIGN: 'ASSIGN',
  UNASSIGN: 'UNASSIGN',
  LOGIN: 'LOGIN',
  LOGOUT: 'LOGOUT'
} as const

export type ACTIVITY_TYPE = (typeof ACTIVITY_TYPE)[keyof typeof ACTIVITY_TYPE]